requirements.txt
matplotlib==3.5.2
numpy==1.21.6
opencv_python==4.7.0.68
pandas==1.3.5
Pillow==9.4.0
progress==1.6
scikit_learn==1.2.2
timm==0.5.4
torch==1.10.1+cu102
torchvision==0.11.2+cu102
